Moved much of the ExtAsync/ExtFuture stuff to attic.
