+++ src/attic/ExtAsync_impl.h
@@ -42,8 +42,8 @@
 // Ours
 #include <utils/DebugHelpers.h>
 //#include "../ExtFuture.h" ///< Can't include this because it uses context_has_event_loop().
-#include "concurrency/ExtFuture_traits.h"
-#include "../ExtAsyncExceptions.h"
+#include "ExtFuture_traits.h"
+#include "ExtAsyncExceptions.h"
 
 //template <class T>
 //class ExtFuture;
