+++ src/attic/ExtFuturePropagationHandler.cpp
