+++ src/attic/ExtFuturePropagationHandler.h
