+++ src/attic/ExtFuture_traits.h
