+++ src/attic/impl/ExtFutureImplHelpers.cpp
