+++ src/attic/impl/ManagedExtFutureWatcher_impl.h
@@ -23,7 +23,7 @@
 #ifndef SRC_CONCURRENCY_IMPL_MANAGEDEXTFUTUREWATCHER_IMPL_H_
 #define SRC_CONCURRENCY_IMPL_MANAGEDEXTFUTUREWATCHER_IMPL_H_
 
-#if 1 // !QT6
+#if 0 // !QT6
 
 // Std C++
 #include <exception>
