+++ src/concurrency/AMLMJob.h
@@ -44,8 +44,7 @@
 #include <future/guideline_helpers.h>
 #include <utils/UniqueIDMixin.h>
 #include "utils/ConnectHelpers.h"
-#include "IExtFutureWatcher.h"
-#include "ExtAsync.h"
+
 
 
 class AMLMJob;
