+++ src/concurrency/CMakeLists.txt
@@ -22,15 +22,8 @@
 # Concurrency files.
 set(HEADER_FILES_UTILS_CONCURRENCY
 	AsyncTaskManager.h
-	ExtFutureState.h
 	ExtFuture.h
 	ExtFutureProgressInfo.h
-	ExtFuturePropagationHandler.h
-	ExtFutureWatcher.h
-	ExtFuture_traits.h
-	IExtFutureWatcher.h
-	ExtAsync.h
-	ExtAsyncExceptions.h
 	AMLMJob.h
 	AMLMJobT.h
 	AMLMCompositeJob.h
@@ -40,28 +33,11 @@ set(HEADER_FILES_UTILS_CONCURRENCY
 	WorkerThreadControllerBase.h
 	)
 set(PRIVATE_HEADER_FILES_UTILS_CONCURRENCY
-	impl/IExecutor.h
-	impl/ExtAsync_impl.h
-	impl/ExtFuture_impl.hpp
-	impl/ExtFuture_make_xxx_future.h
-	impl/ExtFuture_when_all.h
-	impl/ExtFuture_when_any.h
-	impl/ExtFutureImplHelpers.h
-	impl/ExtAsync_RunInThread.h
-	impl/ManagedExtFutureWatcher_impl.h
 	)
 
 set(SOURCE_FILES_UTILS_CONCURRENCY
-	impl/ExtFuture_when_all.cpp
-	impl/ExtFuture_when_any.cpp
-	impl/ExtFutureImplHelpers.cpp
-	impl/ManagedExtFutureWatcher_impl.cpp
-	ExtAsync.cpp
-	ExtAsyncExceptions.cpp
 	ExtFuture.cpp
 	ExtFutureProgressInfo.cpp
-	ExtFuturePropagationHandler.cpp
-	IExtFutureWatcher.cpp
 	AMLMJob.cpp
 	AMLMJobT.cpp
 	AMLMCompositeJob.cpp
