+++ src/logic/jobs/DirectoryScanJob.cpp
@@ -28,7 +28,6 @@
 // Ours
 #include <utils/TheSimplestThings.h>
 #include <logic/DirScanResult.h>
-#include <concurrency/ExtAsync.h>
 #include <utils/Stopwatch.h>
