+++ src/logic/jobs/LibraryRescannerJob.cpp
@@ -29,7 +29,6 @@
 
 // Ours
 #include <utils/DebugHelpers.h>
-#include <concurrency/ExtAsync.h>
 #include <utils/TheSimplestThings.h>
 #include "LibraryModel.h"
